// resources/js/types/TeamProps.ts
export interface TeamProps {
    id: number;
    name: string;
    facebook: string;
    twitter: string;
    linkedin: string;
    title: string;
    image: string;
    email: string;
    status: number;
    about: string;
    role: string;
    email_verified_at: any;
    profile_photo_path: any;
    created_at: string;
    updated_at: string;
}